challenged completed

+++ assets/js/formModal.js
@@ -1,7 +1,30 @@
 $(document).ready(function () {
-  function loadUsers() {
+  function formatDateTime(dateTime, format = "datetime") {
+    if (dateTime.length === 10) {
+      dateTime += "T00:00:00";
+    }
+
+    const dateTimeObj = new Date(dateTime);
+    let formatedDate = "";
+
+    formatedDate += String(dateTimeObj.getDate()).padStart(2, "0") + "/";
+    formatedDate += String(dateTimeObj.getMonth() + 1).padStart(2, "0") + "/";
+    formatedDate += dateTimeObj.getFullYear();
+
+    if (format === "datetime") {
+      formatedDate += " ";
+      formatedDate += String(dateTimeObj.getHours()).padStart(2, "0") + ":";
+      formatedDate += String(dateTimeObj.getMinutes()).padStart(2, "0") + ":";
+      formatedDate += String(dateTimeObj.getSeconds()).padStart(2, "0");
+    }
+
+    return formatedDate;
+  }
+
+  function loadUsers(name = "", email = "") {
     $.ajax({
       url: "load",
+      data: { name: name, email: email },
       method: "GET",
       success: function (response) {
         $("#usersTable tbody").empty();
@@ -11,15 +34,19 @@ $(document).ready(function () {
             <td>${user.id}</td>
             <td>${user.name}</td>
             <td>${user.email}</td>
-            <td>${user.status}</td>
-            <td>${user.admission_date}</td>
-            <td>${user.updated}</td>
-            <td>${user.created}</td>
+            <td>${user.status == 0 ? "Pendente" : "Admitido"}</td>
+            <td>${formatDateTime(user.admission_date, "date")}</td>
+            <td>${formatDateTime(user.updated)}</td>
+            <td>${formatDateTime(user.created)}</td>
             <td>
-              <a class="btn btn-primary button btn-edit-user" data-toggle="modal" data-target="#customModal" data-userid="${user.id}">
+              <a class="btn btn-success button btn-edit-user" data-toggle="modal" data-target="#customModal" data-userid="${
+                user.id
+              }">
                 <span class="iconify" data-icon="mdi:pencil" data-inline="false"></span>
               </a>
-              <a class="btn btn-danger button btn-delete-user" data-userid="${user.id}">
+              <a class="btn btn-danger button btn-delete-user" data-userid="${
+                user.id
+              }">
                 <span class="iconify" data-icon="mdi:trash" data-inline="false"></span>
               </a>
               <a class="btn btn-warning button">
@@ -31,10 +58,22 @@ $(document).ready(function () {
         });
       },
       error: function (xhr, status, error) {
-        console.log(xhr.responseText);
+        Swal.fire("Erro na requisição!", error, "error");
       },
     });
   }
+
+  $("#searchButton").on("click", function () {
+    var name = $("#searchName").val();
+    var email = $("#searchEmail").val();
+    loadUsers(name, email);
+  });
+
+  $("#cleanSearchButton").on("click", function () {
+    $("#searchName").val("");
+    $("#searchEmail").val("");
+    loadUsers();
+  });
 
   $(document).on("click", "#btnAddUser", function () {
     loadUser();
@@ -51,15 +90,29 @@ $(document).ready(function () {
   });
 
   function deleteUser(userId) {
-    $.ajax({
-      url: "delete/" + userId,
-      method: "DELETE",
-      success: function (response) {
-        loadUsers();
-      },
-      error: function (xhr, status, error) {
-        console.log(xhr.responseText);
-      },
+    Swal.fire({
+      title: "Tem certeza?",
+      text: "Você não poderá reverter isso!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Deletar!",
+      cancelButtonText: "Cancelar",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        $.ajax({
+          url: "delete/" + userId,
+          method: "DELETE",
+          success: function (response) {
+            Swal.fire("Deletado!", "O usuário foi deletado.", "success");
+            loadUsers();
+          },
+          error: function (xhr, status, error) {
+            Swal.fire("Erro na requisição!", error, "error");
+          },
+        });
+      }
     });
   }
 
@@ -84,10 +137,15 @@ $(document).ready(function () {
               success: function (response) {
                 $("#customModal").modal("hide");
                 $("#customForm")[0].reset();
+                Swal.fire(
+                  "Cadastrado!",
+                  "O usuário foi cadastrado.",
+                  "success"
+                );
                 loadUsers();
               },
               error: function (xhr, status, error) {
-                console.log(xhr.responseText);
+                Swal.fire("Erro na requisição!", error, "error");
               },
             });
           });
@@ -106,17 +164,18 @@ $(document).ready(function () {
               success: function (response) {
                 $("#customModal").modal("hide");
                 $("#customForm")[0].reset();
+                Swal.fire("Editado!", "O usuário foi editado.", "success");
                 loadUsers();
               },
               error: function (xhr, status, error) {
-                console.log(xhr.responseText);
+                Swal.fire("Erro na requisição!", error, "error");
               },
             });
           });
         }
       },
       error: function (xhr, status, error) {
-        console.log(xhr.responseText);
+        Swal.fire("Erro na requisição!", error, "error");
       },
     });
   }
